Fix negative start coordinates of probes
In case a region of interest is too close to the chromosome beginning,
the first couple of probes covering that region had negative start
start coordinates.

Check the start coordinate of the first probe first and set it to zero
if it is negative (i.e. start tiling the probes from the beginning
of a chromosome).