Fix analysis of TRF overlap using merged probes after TRF filtering
Two separate sets of regions obtained by merging of all probes are now
created to correct this:

- merge of all probes prior to removal of probes overlapping some TRF
- merge of all probes after removal of probes overlapping some TRF

The first is used to calculate a fraction of probes that are overlapped
by some TRF repeat region. The second is used to analyze the total
length of sequence covered by the final set of generated probes.